+++ packages/codemod/src/transforms/app/eslint.ts
@@ -1,0 +1,6 @@
+export default function () {
+  return `const { getESLintConfig } = require('@iceworks/spec');
+
+module.exports = getESLintConfig('rax');
+  `;
+}