+++ packages/plugin-rax-web/src/Plugins/DocumentPlugin.ts
@@ -1,7 +1,8 @@
 import * as path from 'path';
 import * as Module from 'module';
 import * as cheerio from 'cheerio';
-import { registerListenTask } from '../utils/localBuildCache';
+import { getEntriesByRoute } from '@builder/app-helpers';
+import { registerListenTask, getAssets, getEnableStatus, updateEnableStatus } from '../utils/localBuildCache';
 import * as webpackSources from 'webpack-sources';
 import { getInjectedHTML, getBuiltInHtmlTpl, insertCommonElements } from '../utils/htmlStructure';
 
@@ -9,6 +10,7 @@ const PLUGIN_NAME = 'DocumentPlugin';
 const { RawSource } = webpackSources;
 export default class DocumentPlugin {
   options: any;
+  init: boolean;
   constructor(options) {
     this.options = options;
   }
@@ -18,23 +20,31 @@ export default class DocumentPlugin {
       staticConfig,
       api: {
         context: {
-          userConfig: { web = {} },
+          userConfig: { web: webConfig },
+          rootDir,
         },
       },
       documentPath,
       insertScript,
     } = this.options;
+    const { mpa, doctype = '<!DOCTYPE html>', ssr } = webConfig || {};
     // DEF plugin will pass publicPath override compiler publicPath in Weex Type App
     const publicPath = this.options.publicPath || compiler.options.output.publicPath;
-    const doctype = web.doctype || '<!DOCTYPE html>';
     insertCommonElements(staticConfig);
 
     let localBuildTask = registerListenTask();
 
     compiler.hooks.emit.tapAsync(PLUGIN_NAME, async (compilation, callback) => {
-      localBuildTask.then((localBuildAssets) => {
-        // update local build task
-        localBuildTask = registerListenTask();
+      const enableStatus: boolean = getEnableStatus();
+      if (enableStatus) {
+        updateEnableStatus(false);
+        localBuildTask.then(emitAssets).then(() => {
+          localBuildTask = registerListenTask();
+        });
+      } else {
+        emitAssets(getAssets());
+      }
+      function emitAssets(localBuildAssets) {
         const injectedHTML = getInjectedHTML();
         if (insertScript) {
           injectedHTML.scripts.push(`<script>${insertScript}</script>`);
@@ -44,7 +54,8 @@ export default class DocumentPlugin {
           const assets = getAssetsForPage(buildResult, publicPath);
           const title = getTitleByStaticConfig(staticConfig, {
             entryName,
-            mpa: web.mpa,
+            mpa,
+            rootDir,
           });
           let html = '';
           if (documentPath && localBuildAssets[`${entryName}.js`]) {
@@ -55,11 +66,12 @@ export default class DocumentPlugin {
               title,
               pagePath,
             });
+
             const $ = cheerio.load(html, { decodeEntities: false });
-            $('head').append(injectedHTML.scripts);
+            $('#root').after(injectedHTML.scripts);
             html = $.html();
           } else {
-            let initialHTML = '';
+            let initialHTML;
 
             if (localBuildAssets[`${entryName}.js`]) {
               const bundleContent = localBuildAssets[`${entryName}.js`].source();
@@ -75,30 +87,25 @@ export default class DocumentPlugin {
               initialHTML,
               spmA: staticConfig.spm,
               spmB: spm,
-            });
+            }, ssr);
           }
 
           compilation.assets[`${entryName}.html`] = new RawSource(html);
         });
 
         callback();
-      });
+      }
     });
   }
 }
 
-function getTitleByStaticConfig(staticConfig, { entryName, mpa }): string {
+function getTitleByStaticConfig(staticConfig, { entryName, mpa, rootDir }): string {
   if (!mpa) return staticConfig.window?.title;
-  const route = staticConfig.routes.find(({ source, name }) => {
-    let pageEntry;
-    if (name) {
-      pageEntry = name;
-    } else if (source) {
-      const dir = path.dirname(source);
-      pageEntry = path.parse(dir).name.toLocaleLowerCase();
-    }
-    return pageEntry === entryName;
-  });
+  const route = staticConfig.routes
+    .reduce((prev, curr) => {
+      return [...prev, ...getEntriesByRoute(curr, rootDir)];
+    }, [])
+    .find(({ entryName: pageEntry }) => pageEntry === entryName);
   return route.window?.title || staticConfig.window?.title;
 }