+++ packages/plugin-rax-web/src/__tests__/htmlStructure.test.ts
@@ -47,7 +47,7 @@ describe('generate html structure', () => {
         },
         spmA: 'a1234',
         spmB: 'b1234',
-      }),
+      }, true),
     ).toEqual(`
   <!DOCTYPE html>
   <html>
@@ -66,11 +66,13 @@ describe('generate html structure', () => {
 
     </head>
     <body data-spm="b1234">
-      <div id=\"root\"></div>
+      <!--__BEFORE_ROOT__-->
+      <div id=\"root\"><!--__INNER_ROOT__--></div>
       <script src=\"https://g.alicdn.com/code/lib/rax/1.1.4/rax.min.js\"></script>
 <script src=\"https://g.alicdn.com/code/lib/react/17.0.0/react.min.js\"></script>
 <script src=\"https://g.alicdn.com/ali-lib/appear-polyfill/0.1.2/index.js\" ></script>
 
+      <!--__AFTER_ROOT__-->
       <script crossorigin=\"anonymous\" type=\"application/javascript\" src=\"/home.js\"></script>
 
     </body>