+++ packages/rax-webpack-config/src/setWebpackLoaders.js
@@ -93,7 +93,7 @@ module.exports = (config, { rootDir, mode, babelConfig }) => {
   [
     ['css'],
     ['scss', [['sass-loader', require.resolve('sass-loader')]]],
-    ['less', [['less-loader', require.resolve('less-loader'), { javascriptEnabled: true }]]],
+    ['less', [['less-loader', require.resolve('less-loader'), { lessOptions: { javascriptEnabled: true } }]]],
   ].forEach(([style, loaders]) => {
     configCSSRule(config, style, mode, loaders || []);
   });